release

+++ src/main/java/controller/AdminCategoryPropertyController.java
@@ -3,7 +3,10 @@ package controller;
 import com.google.gson.Gson;
 import dao.CategoryPropertyDAO;
 import entity.CategoryProperty;
+import entity.User;
 import java.util.List;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import org.springframework.http.*;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
@@ -13,60 +16,92 @@ import org.json.JSONObject;
 @Controller
 public class AdminCategoryPropertyController {
 
+    HttpHeaders responseHeaders = new HttpHeaders();
     // @Autowired
     private CategoryPropertyDAO CategoryPropertyDAO = new CategoryPropertyDAO();
 
-    public String index(ModelMap cate) {
-        return "/admin/categoryProperty";
+    public String index(ModelMap cate, HttpServletRequest req) {
+        HttpSession session = req.getSession();
+
+        if (session.getAttribute("user") == null) {
+            return "redirect:/";
+        }
+
+        User user = (User) session.getAttribute("user");
+
+        if (user.getUserGroup() == 0) {
+            return "/admin/categoryProperty";
+        } else {
+            return "redirect:/";
+        }
     }
 
     // @RequestMapping(value = "/json", method = RequestMethod.GET, produces = "application/json")
-    public ResponseEntity<String> getlist() {
-        JSONObject jsonOB = new JSONObject();
-        try {
-            List<CategoryProperty> catlist = CategoryPropertyDAO.getlist();
+    public ResponseEntity<String> getlist(HttpServletRequest req) {
 
-            String json = new Gson().toJson(catlist);
-            if (json != null) {
-                jsonOB.put("result", json);
-                jsonOB.put("message", "success_ok");
-            } else {
+        HttpSession SESSION = req.getSession();
+
+        JSONObject jsonOB = new JSONObject();
+
+        if (SESSION.getAttribute("user") == null) {
+            jsonOB.put("message", "Out of session!");
+        } else {
+            try {
+                List<CategoryProperty> catlist = CategoryPropertyDAO.getlist();
+
+                String json = new Gson().toJson(catlist);
+                if (json != null) {
+                    jsonOB.put("result", json);
+                    jsonOB.put("message", "success_ok");
+                } else {
+                    jsonOB.put("message", "success_fail");
+                }
+            } catch (Exception e) {
                 jsonOB.put("message", "success_fail");
             }
-        } catch (Exception e) {
-            jsonOB.put("message", "success_fail");
         }
 
         String json1 = new Gson().toJson(jsonOB);
-        return new ResponseEntity<String>(json1, HttpStatus.CREATED);
+        responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+        return new ResponseEntity<String>(json1, responseHeaders, HttpStatus.CREATED);
     }
 
-    public ResponseEntity<String> getitemdetail(@RequestParam("id") String catID) {
-        JSONObject jsonOB = new JSONObject();
-        try {
-            List<CategoryProperty> catlist = CategoryPropertyDAO.getitemDetail(Integer.parseInt(catID));
+    public ResponseEntity<String> getitemdetail(HttpServletRequest req, @RequestParam("id") String catID) {
 
-            String json = new Gson().toJson(catlist);
-            if (json != "null") {
-                jsonOB.put("result", json);
-                jsonOB.put("message", "success_ok");
-            } else {
+        HttpSession SESSION = req.getSession();
+
+        JSONObject jsonOB = new JSONObject();
+
+        if (SESSION.getAttribute("user") == null) {
+            jsonOB.put("message", "Out of session!");
+        } else {
+            try {
+                List<CategoryProperty> catlist = CategoryPropertyDAO.getitemDetail(Integer.parseInt(catID));
+
+                String json = new Gson().toJson(catlist);
+                if (json != "null") {
+                    jsonOB.put("result", json);
+                    jsonOB.put("message", "success_ok");
+                } else {
+                    jsonOB.put("message", "success_fail");
+                }
+            } catch (Exception e) {
                 jsonOB.put("message", "success_fail");
             }
-        } catch (Exception e) {
-            jsonOB.put("message", "success_fail");
         }
 
         String json1 = new Gson().toJson(jsonOB);
-        return new ResponseEntity<String>(json1, HttpStatus.CREATED);
+        responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+        return new ResponseEntity<String>(json1, responseHeaders, HttpStatus.CREATED);
     }
 
-    public ResponseEntity<String> insert(
+    public ResponseEntity<String> insert(HttpServletRequest req,
             @RequestParam("id") String id,
             @RequestParam("catId") String catId,
             @RequestParam("name") String name,
             @RequestParam("status") String status,
             ModelMap cate) {
+        HttpSession SESSION = req.getSession();
 
         JSONObject jsonOB = new JSONObject();
 
@@ -78,22 +113,29 @@ public class AdminCategoryPropertyController {
 
         }
 
-        try {
-            if (CategoryPropertyDAO.insert(idTemp, catId, name, status)) {
-                jsonOB.put("message", "success_ok");
-            } else {
+        if (SESSION.getAttribute("user") == null) {
+            jsonOB.put("message", "Out of session!");
+        } else {
+            try {
+                if (CategoryPropertyDAO.insert(idTemp, catId, name, status)) {
+                    jsonOB.put("message", "success_ok");
+                } else {
+                    jsonOB.put("message", "success_fail");
+                }
+            } catch (Exception e) {
                 jsonOB.put("message", "success_fail");
             }
-        } catch (Exception e) {
-            jsonOB.put("message", "success_fail");
         }
 
         String json = new Gson().toJson(jsonOB);
-        return new ResponseEntity<String>(json, HttpStatus.CREATED);
+        responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+        return new ResponseEntity<String>(json, responseHeaders, HttpStatus.CREATED);
     }
 
-    public ResponseEntity<String> delete(
+    public ResponseEntity<String> delete(HttpServletRequest req,
             @RequestParam("id") String id) {
+
+        HttpSession SESSION = req.getSession();
 
         JSONObject jsonOB = new JSONObject();
 
@@ -105,17 +147,22 @@ public class AdminCategoryPropertyController {
 
         }
 
-        try {
-            if (CategoryPropertyDAO.delete(idTemp)) {
-                jsonOB.put("message", "success_ok");
-            } else {
+        if (SESSION.getAttribute("user") == null) {
+            jsonOB.put("message", "Out of session!");
+        } else {
+            try {
+                if (CategoryPropertyDAO.delete(idTemp)) {
+                    jsonOB.put("message", "success_ok");
+                } else {
+                    jsonOB.put("message", "success_fail");
+                }
+            } catch (Exception e) {
                 jsonOB.put("message", "success_fail");
             }
-        } catch (Exception e) {
-            jsonOB.put("message", "success_fail");
         }
 
         String json = new Gson().toJson(jsonOB);
-        return new ResponseEntity<String>(json, HttpStatus.CREATED);
+        responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+        return new ResponseEntity<String>(json, responseHeaders, HttpStatus.CREATED);
     }
 }
